// frontend/src/types/generated/ProviderControl.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProviderControlOption } from "./ProviderControlOption.js";

export type ProviderControl = { key: string, label: string, kind: string, options: Array<ProviderControlOption>, default_value: string, };